+++ common/dbl.hh
@@ -60,15 +60,15 @@ namespace csl
 
       public:
        typedef double value_t;
-       enum { var_type_v =  CSL_TYPE_DOUBLE }; 
-      
+       enum { var_type_v =  CSL_TYPE_DOUBLE };
+
        inline dbl() : var(), value_(0.0) { }             ///<default constructor
        inline dbl(double v) : var(), value_(v) { }       ///<copy constructor
        virtual inline ~dbl() {}                          ///<destructor
        inline value_t value() const { return value_; }    ///<returns the double value
        inline int var_type() const { return var_type_v; } ///<value type helps distinguish from other var types
        inline void reset() { value_ = 0.0; }             ///<resets the internal value to 0.0
-       
+
         /** @brief copy operator */
         inline dbl & operator=(const dbl & other)
         {
@@ -78,10 +78,10 @@ namespace csl
 
         /** @brief returns a const pointer to internal data */
         inline const unsigned char * ucharp_data() const { return reinterpret_cast<const unsigned char *>(&value_); }
-        
+
         /** @brief returns the size of the variable data */
-        inline size_t var_size() const { return sizeof(value_); }
-        
+        inline uint64_t var_size() const { return sizeof(value_); }
+
         /* conversions to other types */
         /**
         @brief convert to common::int64
@@ -166,7 +166,7 @@ namespace csl
         this function assumes that v has enough space. sz will be updated to indicate, how
         many bytes were stored.
          */
-        bool to_binary(unsigned char * v, size_t & sz) const;
+        bool to_binary(unsigned char * v, uint64_t & sz) const;
 
         /**
         @brief convert to (ptr+size)
@@ -178,7 +178,7 @@ namespace csl
         many bytes were stored.
 
          */
-        bool to_binary(void * v, size_t & sz) const;
+        bool to_binary(void * v, uint64_t & sz) const;
 
         /** @todo not implemented */
         bool to_xdr(xdrbuf & b) const;
@@ -289,7 +289,7 @@ namespace csl
 
         this function delegates the conversion task to the common::binry class
          */
-        bool from_binary(const unsigned char * v,size_t sz);
+        bool from_binary(const unsigned char * v,uint64_t sz);
 
         /**
         @brief convert a memory region
@@ -298,7 +298,7 @@ namespace csl
 
         this function delegates the conversion task to the common::binry class
          */
-        bool from_binary(const void * v,size_t sz);
+        bool from_binary(const void * v,uint64_t sz);
 
         /** @todo not implemented */
         bool from_xdr(xdrbuf & v);
